revise z-checker-installer.sh

+++ z-checker-install.sh
@@ -230,8 +230,11 @@ modifyZCConfig ./zc.config checkingStatus PROBE_COMPRESSOR
 #---------- download MGARD and set the configuration ------------
 cd $rootDir/mgard-patches
 ./compile-mgard-zchecker.sh
-cp ../libpressio/test/mgardfloat_CompDecomp ../MGARD/build/bin
-cp ../libpressio/test/mgarddouble_CompDecomp ../MGARD/build/bin
+if [ -f ../libpressio/test/mgardfloat_CompDecomp ]
+then
+	cp ../libpressio/test/mgardfloat_CompDecomp ../MGARD/build/bin
+	cp ../libpressio/test/mgarddouble_CompDecomp ../MGARD/build/bin
+fi
 cd ..
 
 #---------- download bit_grooming and set the configuration -----------
